+++ tests/Fixtures/ExampleEntity.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types=1);
+
+namespace Tests\Torr\Rad\Fixtures;
+
+use Torr\Rad\Entity\Interfaces\EntityInterface;
+use Torr\Rad\Entity\Traits\IdTrait;
+use Torr\Rad\Entity\Traits\TimestampsTrait;
+
+/**
+ */
+class ExampleEntity implements EntityInterface
+{
+	use IdTrait;
+	use TimestampsTrait;
+}