+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 *   (feature) Add `DataContainer`.
 *   (feature) Add Twig function: `data_container`
 *   (feature) Add Twig filter: `appendToArrayKey`
+*   (feature) Add `Model` + `ModelInterface`.